login loading stage for dispatcher app
