+++ app/login/index/controller.js
@@ -2,6 +2,6 @@ import Ember from 'ember';
 
 export default Ember.Controller.extend({
 
-  invalidCredentials: false
+  invalidCredentials: false,
 
 });
