+++ app/login/index/template.hbs
@@ -1,0 +1,1 @@
+{{login-form login="login" forgotPassword="forgotPassword" displayError=invalidCredentials loading=loading}}
