+++ app/login/loading/template.hbs
@@ -1,0 +1,1 @@
+{{login-form loading=true}}
