--- app/login/template.hbs
@@ -1,1 +1,0 @@
-{{login-form login="login" forgotPassword="forgotPassword" displayError=invalidCredentials}}
